updated all api's by including the session to handle authorization

+++ src/routes/api/tags/[id]/+server.ts
@@ -1,16 +1,23 @@
 import { json } from "@sveltejs/kit";
 import prisma from "../../../../../prisma/client";
 
-
-export async function DELETE({request, params}:{request:Request, params:{id:number}}){
-    try{
-        await prisma.tag.delete({
-            where:{
-                id:Number(params.id)
-            },
-        });
-        return json({message:"Tag deleted successfully!"},{status:201})
-    }catch (error:any) {
+export async function DELETE({
+  request,
+  params: { id },
+  locals: { getSession },
+}) {
+  const session = await getSession();
+  try {
+    if (session) {
+      await prisma.tag.delete({
+        where: {
+          id: Number(id),
+        },
+      });
+      return json({ message: "Tag deleted successfully!" }, { status: 201 });
+    }
+    return json({ message: "Unauthorized" }, { status: 401 });
+  } catch (error: any) {
     return json(
       { message: "An server error has occurred", error: error.message },
       { status: 500 }
